third_party/blink/web_tests/http/tests/security/script-crossorigin-redirect-no-cors-expected.txt

CONSOLE ERROR: Access to script at 'http://localhost:8000/security/resources/cors-redirect.php?mode=no&url=http://localhost:8000/security/resources/localScript.js' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to script at 'http://localhost:8000/security/resources/cors-redirect.php?mode=no&url=http://localhost:8000/security/resources/script-allow-star.php' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to script at 'http://localhost:8000/security/resources/cors-redirect.php?mode=no&url=http://localhost:8000/security/resources/script-allow-credentials.php' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to script at 'http://localhost:8000/security/resources/cors-redirect.php?mode=no&url=http://127.0.0.1:8000/security/resources/script-allow-credentials.php' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Testing the handling of CORS-enabled

PASS/FAIL descriptions are of the form, 'CORS request type': 'redirect CORS type' => 'resource'


PASS: Anonymous request: no-CORS => no-CORS script resource.
PASS: Anonymous request: no-CORS => anonymous-CORS script resource.
PASS: Credentialled request: no-CORS => credential-CORS script resource (same origin.)
PASS: Credentialled request: no-CORS => credential-CORS script resource (cross origin.)
PASS successfullyParsed is true

TEST COMPLETE
